+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 1. [](#new)
     * `FormInterface` now implements `RenderInterface` 
     * Added new `FormInterface::getTask()` method which reads the task from `form.task` in the blueprint 
-
+1. [](#bugfix)
+    * Fixed bug in text field filtering: return empty string if value isn't a string or number
+    
 # v1.6.5
 ## 04/15/2019