+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 1. [](#new)
     * `FormInterface` now implements `RenderInterface` 
     * Added new `FormInterface::getTask()` method which reads the task from `form.task` in the blueprint 
-
+1. [](#bugfix)
+    * Rollback `redirect_default_route` logic as it has issues with multilang [#2459](https://github.com/getgrav/grav/issues/2459)
+    
 # v1.6.5
 ## 04/15/2019