+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.6.6
-## mm/dd/2019
+## 04/17/2019
 
 1. [](#new)
     * `FormInterface` now implements `RenderInterface`